requirements.txt
Django>=4.2
celery>=5.3
redis
djangorestframework
gunicorn
